+++ storybook/package.json
@@ -25,8 +25,8 @@
   "scripts": {
     "_lint": "eslint -c ../superset-frontend/.eslintrc.js --ext .js,.jsx,.ts,tsx .",
     "_prettier": "prettier --ignore-path=../superset-frontend/.prettierignore './({src,spec,cypress-base,plugins,packages,.storybook}/**/*{.js,.jsx,.ts,.tsx,.css,.less,.scss,.sass}|package.json)'",
-    "build": "storybook build",
-    "dev": "storybook dev -p 9001",
+    "build": "build-storybook",
+    "dev": "start-storybook -p 9001",
     "format": "npm run _prettier -- --write",
     "lint": "npm run _lint && npm run type",
     "lint-fix": "npm run _lint -- --fix && npm run type",
@@ -43,17 +43,16 @@
     "@emotion/react": "11.4.1",
     "@emotion/styled": "11.3.0",
     "@react-icons/all-files": "^4.1.0",
-    "@storybook/addon-actions": "^7.5.1",
-    "@storybook/addon-docs": "^7.5.3",
-    "@storybook/addon-essentials": "^7.5.3",
-    "@storybook/addon-knobs": "^7.0.2",
-    "@storybook/addon-links": "^7.5.1",
-    "@storybook/builder-webpack5": "^7.5.3",
-    "@storybook/client-api": "^7.5.3",
-    "@storybook/manager-api": "^7.5.3",
-    "@storybook/preview-api": "^7.5.3",
-    "@storybook/react": "^7.5.3",
-    "@storybook/react-webpack5": "^7.5.1",
+    "@storybook/addon-actions": "^6.3.12",
+    "@storybook/addon-knobs": "^6.3.1",
+    "@storybook/addon-links": "^6.3.12",
+    "@storybook/addons": "^6.3.12",
+    "@storybook/react": "^6.3.12",
+    "@storybook/addon-docs": "^6.5.10",
+    "@storybook/addon-essentials": "^6.4.22",
+    "@storybook/builder-webpack5": "^6.4.22",
+    "@storybook/client-api": "^6.4.22",
+    "@storybook/manager-webpack5": "^6.4.22",
     "@superset-ui/core": "*",
     "@superset-ui/custom-plugin-chart-hello-world": "file:../custom-frontend/plugins/custom-plugin-chart-hello-world",
     "@types/react-loadable": "^5.5.3",
@@ -68,9 +67,8 @@
     "react-hot-loader": "^4.13.1",
     "react-loadable": "^5.5.0",
     "react-resizable": "^3.0.4",
-    "storybook": "^7.5.3",
+    "storybook": "^6.3.12",
     "storybook-addon-jsx": "^7.3.14",
-    "storybook-addon-paddings": "^5.0.0",
     "yargs": "^15.4.1"
   },
   "devDependencies": {
